pyproject.toml
[project]
name = "arifos"
version = "2026.04.24"
description = "arifOS v2026.04.24-KANON — Constitutional AI Governance System. 13 canonical tools, 13 constitutional floors (F1-F13), streamable-http MCP transport. CANONICAL SOURCE OF TRUTH: arifOS repository."
readme = "README.md"
requires-python = ">=3.12"
license = "AGPL-3.0-only"
license-files = ["LICENSE"]
keywords = [
    "ai-governance",
    "constitutional-ai",
    "mcp-server",
    "trinity-architecture",
    "agi-asi-apex",
]
authors = [{ name = "Muhammad Arif bin Fazil", email = "arifbfazil@gmail.com" }]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
    "Typing :: Typed",
]

dependencies = [
    "fastmcp>=3.2.0",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sse-starlette>=1.8.2",
    "starlette>=0.30.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "anyio>=4.0.0",
    "python-dotenv>=1.0.0",
    "prometheus-client>=0.19.0",
    "rich>=13.7.0",
    "psutil>=5.9.0",
    "numpy>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
arifos-mcp = "arifosmcp.server:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["arifosmcp*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
strict_optional = true

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
